Module split and rename suggestions
