+++ cardano-cli/cardano-cli.cabal
@@ -116,6 +116,8 @@ library
     Cardano.CLI.EraBased.Run.StakePool
     Cardano.CLI.EraBased.Run.TextView
     Cardano.CLI.EraBased.Run.Transaction
+    Cardano.CLI.EraBased.Script.Mint.Read
+    Cardano.CLI.EraBased.Script.Mint.Types
     Cardano.CLI.EraBased.Transaction.HashCheck
     Cardano.CLI.Helpers
     Cardano.CLI.IO.Lazy
@@ -135,7 +137,6 @@ library
     Cardano.CLI.Options.Key
     Cardano.CLI.Options.Node
     Cardano.CLI.Options.Ping
-    Cardano.CLI.Plutus.Minting
     Cardano.CLI.Orphans
     Cardano.CLI.Parser
     Cardano.CLI.Read
