+++ cardano-cli/src/Cardano/CLI/EraBased/Commands/Transaction.hs
@@ -25,7 +25,7 @@ import qualified Cardano.Api.Experimental as Exp
 import           Cardano.Api.Ledger (Coin)
 import           Cardano.Api.Shelley
 
-import           Cardano.CLI.Plutus.Minting
+import           Cardano.CLI.EraBased.Script.Mint.Types
 import           Cardano.CLI.Types.Common
 import           Cardano.CLI.Types.Governance
