+++ cardano-cli/src/Cardano/CLI/EraBased/Script/Mint/Read.hs
@@ -1,106 +1,17 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE GADTs #-}
-{-# LANGUAGE LambdaCase #-}
 
-module Cardano.CLI.Plutus.Minting
-  ( CliMintScriptRequirements (..)
-  , MintScriptWitnessWithPolicyId (..)
-  , createSimpleOrPlutusScriptFromCliArgs
-  , createSimpleReferenceScriptFromCliArgs
-  , createPlutusReferenceScriptFromCliArgs
-  , CliScriptWitnessError
-  , readMintScriptWitness
+module Cardano.CLI.EraBased.Script.Mint.Read
+  ( readMintScriptWitness
   )
 where
 
 import           Cardano.Api
 import           Cardano.Api.Shelley
 
+import           Cardano.CLI.EraBased.Script.Mint.Types
 import           Cardano.CLI.Read
-import           Cardano.CLI.Types.Common (ScriptDataOrFile)
-
--- We always need the policy id when constructing a transaction that mints.
--- In the case of reference scripts, the user currently must provide the policy id (script hash)
--- in order to correctly construct the transaction.
-data MintScriptWitnessWithPolicyId era
-  = MintScriptWitnessWithPolicyId
-  { mswPolId :: PolicyId
-  , mswScriptWitness :: ScriptWitness WitCtxMint era
-  }
-  deriving Show
-
-data CliMintScriptRequirements
-  = OnDiskSimpleOrPlutusScript OnDiskSimpleOrPlutusScriptCliArgs
-  | OnDiskSimpleRefScript SimpleRefScriptCliArgs
-  | OnDiskPlutusRefScript PlutusRefScriptCliArgs
-  deriving Show
-
-data OnDiskSimpleOrPlutusScriptCliArgs
-  = OnDiskSimpleScriptCliArgs
-      (File ScriptInAnyLang In)
-  | OnDiskPlutusScriptCliArgs
-      (File ScriptInAnyLang In)
-      ScriptDataOrFile
-      ExecutionUnits
-  deriving Show
-
-createSimpleOrPlutusScriptFromCliArgs
-  :: File ScriptInAnyLang In
-  -> Maybe (ScriptDataOrFile, ExecutionUnits)
-  -> CliMintScriptRequirements
-createSimpleOrPlutusScriptFromCliArgs scriptFp Nothing =
-  OnDiskSimpleOrPlutusScript $ OnDiskSimpleScriptCliArgs scriptFp
-createSimpleOrPlutusScriptFromCliArgs scriptFp (Just (redeemerFile, execUnits)) =
-  OnDiskSimpleOrPlutusScript $ OnDiskPlutusScriptCliArgs scriptFp redeemerFile execUnits
-
-data SimpleRefScriptCliArgs
-  = SimpleRefScriptCliArgs
-      TxIn
-      PolicyId
-  deriving Show
-
-createSimpleReferenceScriptFromCliArgs
-  :: TxIn
-  -> PolicyId
-  -> CliMintScriptRequirements
-createSimpleReferenceScriptFromCliArgs txin polid =
-  OnDiskSimpleRefScript $ SimpleRefScriptCliArgs txin polid
-
-data PlutusRefScriptCliArgs
-  = PlutusRefScriptCliArgs
-      TxIn
-      AnyPlutusScriptVersion
-      ScriptDataOrFile
-      ExecutionUnits
-      PolicyId
-  deriving Show
-
-createPlutusReferenceScriptFromCliArgs
-  :: TxIn
-  -> AnyPlutusScriptVersion
-  -> ScriptDataOrFile
-  -> ExecutionUnits
-  -> PolicyId
-  -> CliMintScriptRequirements
-createPlutusReferenceScriptFromCliArgs txin scriptVersion scriptData execUnits polid =
-  OnDiskPlutusRefScript $ PlutusRefScriptCliArgs txin scriptVersion scriptData execUnits polid
-
-data CliScriptWitnessError
-  = SimpleScriptWitnessDecodeError ScriptDecodeError
-  | PlutusScriptWitnessDecodeError PlutusScriptDecodeError
-  | PlutusScriptWitnessLanguageNotSupportedInEra
-      AnyPlutusScriptVersion
-      AnyShelleyBasedEra
-  | PlutusScriptWitnessRedeemerError ScriptDataError
-
-instance Error CliScriptWitnessError where
-  prettyError = \case
-    SimpleScriptWitnessDecodeError err -> prettyError err
-    PlutusScriptWitnessDecodeError err -> prettyError err
-    PlutusScriptWitnessLanguageNotSupportedInEra version era ->
-      "Plutus script version " <> pshow version <> " is not supported in era " <> pshow era
-    PlutusScriptWitnessRedeemerError err -> renderScriptDataError err
 
 readMintScriptWitness
   :: MonadIOTransError (FileError CliScriptWitnessError) t m
@@ -190,6 +101,7 @@ readMintScriptWitness
               redeemer
               execUnits
 
+-- TODO: Remove me when exposed from cardano-api
 sbeToSimpleScriptLangInEra
   :: ShelleyBasedEra era -> ScriptLanguageInEra SimpleScript' era
 sbeToSimpleScriptLangInEra ShelleyBasedEraShelley = SimpleScriptInShelley
